More tps schenanigans

+++ Common/Models/WebSocket/LCG/TpsData.cs
@@ -2,6 +2,6 @@
 
 public sealed class TpsData
 {
-    public required float Client { get; set; }
-    public required float Server { get; set; }
+    public required byte Client { get; set; }
+    public required byte Server { get; set; }
 }
